pyproject.toml
[tool.poetry]
name = "reconlib"
version = "0.2.0"
description = "A collection of modules and helpers for active and passive reconnaissance of remote hosts"
authors = ["eonraider <livewire_voodoo@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
